refactor: modularize workspace service into specialized modules
- Extract CRUD operations into workspace-crud.ts
- Extract access/permission logic into workspace-access.ts
- Extract validation logic into workspace-validation.ts
- Create WorkspaceService orchestration class
- Maintain 100% backward compatibility with existing imports
- Add comprehensive documentation and usage examples
- Improve type safety and remove any type usage
- Enhance validation with name and description checks
